requirements.txt
ultralytics==8.0.226
clearml==1.13.2
albumentations==1.3.1
onnxruntime==1.16.3
onnx==1.15.0
streamlit==1.29.0
Pillow==10.1.0
moviepy==1.0.3